+++ src/main/java/common/typeDouble.java
@@ -16,7 +16,7 @@ public class typeDouble implements Type {
     }
 
     @Override
-    public long getDataSize() {
+    public int getDataSize() {
         return 8;
     }
