+++ src/main/java/common/typeLong.java
@@ -17,7 +17,7 @@ public class typeLong implements Type {
     }
 
     @Override
-    public long getDataSize() {
+    public int getDataSize() {
         return 8;
     }
